Fixed spelling mistakes in the comments of certain tests. Renamed the MathUtils class to Math. Created tests for the Math class. Added the tests for the isPowerOfTwo Method.
